fixed few more namespace usage and cast

+++ examples/lighting-app/nxp/k32w/k32w0/main/AppTask.cpp
@@ -809,7 +809,7 @@ void AppTask::OnTriggerEffect(Identify * identify)
 
     case Clusters::Identify::EffectIdentifierEnum::kChannelChange:
         ChipLogProgress(Zcl, "Channel Change effect not supported, using effect %d",
-                        Clusters::Identify::EffectIdentifierEnum::kBlink);
+                        (uint8_t) Clusters::Identify::EffectIdentifierEnum::kBlink);
         timerDelay = 2;
         break;
